--- tests/test_eeg.py
@@ -1,26 +1,0 @@
-import pytest
-import doctest
-import os
-import numpy as np
-import pandas as pd
-import neurokit as nk
-
-
-run_tests_in_local = False
-
-
-#==============================================================================
-# data
-#==============================================================================
-#def test_read_acqknowledge():
-#
-#    assert 3 == 3
-
-
-
-
-if __name__ == '__main__':
-#    nose.run(defaultTest=__name__)
-    doctest.testmod()
-    pytest.main()
-
